api2/boerse_frankfurt_rest/app/requirements.txt
pymongo
Flask-Graphite